+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## Unreleased
 
-## 💡 Enhancements 💡
+## 🛑 Breaking changes 🛑
 
+- Remove `config.NewConfigMapFrom[File|Buffer]`, add testonly version (#4502)
 - `configtls`: TLS 1.2 is the new default mininum version (#4503)
 
 ## 🧰 Bug fixes 🧰